Low memory merging on XMM-LSS
On XMM-LSS #4 we have a memory issue. I am trying to perform the merge without the photometry and then add the photometry at the end. If this works we will probably have to have separate merging notebooks for the surveys with overlapping bands which include the choices i.e. merge uds and dxs in to a single UKIDSS table before the final merge. If the merge works with all columsn apart from those starting with f_, feer_, m_, or merr_ this method should work.