Added DES on all relevant fields
I havent added it to merge sections yet. I presume this will add a lot to some fields so I should add it to any fields not yet run. In fact I think it makes sense to add it to all fields but for fields already run through XID+ of photo-z etc it should not be the DR1 masterlist #4 #6 #11 #15 #17 #18 #20